dodanie wczytywania parametrów z pliku
- predkosc wezy,
- czestotliwosc pojawiania sie bonusus,
- ilosc zyc wezy.

+++ SNAKE_class.h
@@ -36,10 +36,9 @@ public:
     int control;
     int lenght;
     int direction;
-    int lifes;
+    float lifes;
     bool direction_changed;
 
 private:
-
 };
 #endif // SNAKE_CLASS_H
